FM-99: Add tm module
