+++ fendermint/eth/api/src/apis/eth.rs
@@ -12,91 +12,9 @@ use fendermint_vm_actor_interface::eam::EthAddress;
 use fendermint_vm_core::chainid;
 use fvm_shared::{address::Address, error::ExitCode};
 use jsonrpc_v2::{ErrorLike, Params};
-use tendermint::block::Height;
-use tendermint_rpc::{
-    endpoint::{block, commit, header, header_by_hash},
-    Client,
-};
+use tendermint_rpc::{endpoint::block, Client};
 
-use crate::{JsonRpcData, JsonRpcResult};
-
-/// Get the Tendermint block at a specific height.
-async fn tm_block_by_height<C>(
-    client: &C,
-    block_number: ethtypes::BlockNumber,
-) -> JsonRpcResult<tendermint::Block>
-where
-    C: Client + Sync,
-{
-    let block = match block_number {
-        ethtypes::BlockNumber::Number(height) => {
-            let height = Height::try_from(height.as_u64())?;
-            let res: block::Response = client.block(height).await?;
-            res.block
-        }
-        ethtypes::BlockNumber::Finalized
-        | ethtypes::BlockNumber::Latest
-        | ethtypes::BlockNumber::Safe
-        | ethtypes::BlockNumber::Pending => {
-            let res: block::Response = client.latest_block().await?;
-            res.block
-        }
-        ethtypes::BlockNumber::Earliest => {
-            let res: block::Response = client.block(Height::from(1u32)).await?;
-            res.block
-        }
-    };
-    Ok(block)
-}
-
-/// Get the Tendermint header at a specific height.
-async fn tm_header_by_height<C>(
-    client: &C,
-    block_number: ethtypes::BlockNumber,
-) -> JsonRpcResult<tendermint::block::Header>
-where
-    C: Client + Sync,
-{
-    let header = match block_number {
-        ethtypes::BlockNumber::Number(height) => {
-            let height = Height::try_from(height.as_u64())?;
-            let res: header::Response = client.header(height).await?;
-            res.header
-        }
-        ethtypes::BlockNumber::Finalized
-        | ethtypes::BlockNumber::Latest
-        | ethtypes::BlockNumber::Safe
-        | ethtypes::BlockNumber::Pending => {
-            let res: commit::Response = client.latest_commit().await?;
-            res.signed_header.header
-        }
-        ethtypes::BlockNumber::Earliest => {
-            let res: header::Response = client.header(Height::from(1u32)).await?;
-            res.header
-        }
-    };
-    Ok(header)
-}
-
-/// Get the Tendermint header by hash
-async fn tm_header_by_hash<C>(
-    client: &C,
-    block_hash: ethtypes::H256,
-) -> JsonRpcResult<tendermint::block::Header>
-where
-    C: Client + Sync,
-{
-    let hash = tendermint::Hash::Sha256(*block_hash.as_fixed_bytes());
-    let res: header_by_hash::Response = client.header_by_hash(hash).await?;
-    match res.header {
-        Some(header) => Ok(header),
-        None => Err(jsonrpc_v2::Error::Full {
-            code: ExitCode::USR_NOT_FOUND.code(),
-            message: format!("block {block_hash} not found"),
-            data: None,
-        }),
-    }
-}
+use crate::{tm, JsonRpcData, JsonRpcResult};
 
 /// Returns a list of addresses owned by client.
 ///
@@ -140,8 +58,8 @@ where
     C: Client + Sync + Send,
 {
     let header = match block_id {
-        BlockId::Number(n) => tm_header_by_height(data.client.underlying(), n).await?,
-        BlockId::Hash(h) => tm_header_by_hash(data.client.underlying(), h).await?,
+        BlockId::Number(n) => tm::header_by_height(data.client.underlying(), n).await?,
+        BlockId::Hash(h) => tm::header_by_hash(data.client.underlying(), h).await?,
     };
     let height = header.height;
     let addr = Address::from(&EthAddress(addr.0));
@@ -172,7 +90,7 @@ pub async fn get_block_transaction_count_by_number<C: Client>(
 where
     C: Client + Sync,
 {
-    let block = tm_block_by_height(data.client.underlying(), block_number).await?;
+    let block = tm::block_by_height(data.client.underlying(), block_number).await?;
 
     Ok(ethtypes::U64::from(block.data.len()))
 }
